Fixed getting latency metrics (#1855)
